requirements.txt
parliament-functions==0.1.0
yagmail
